0604 JWT 주문, 도서 API에 추가

+++ controller/CartController.js
@@ -1,7 +1,7 @@
+const ensureAuthorization = require('../auth') // 인증 모듈
 const jwt = require("jsonwebtoken");
 const conn = require('../mariadb'); // db 모듈
 const {StatusCodes} = require('http-status-codes'); // status code 모듈
-const dotenv = require('dotenv');
 
 // 장바구니 담기
 const addToCart = (req, res) => {
@@ -51,51 +51,53 @@ const getCartItems = (req, res) => {
         let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price 
         FROM cartItems LEFT JOIN books 
         ON cartItems.book_id = books.id
-        WHERE user_id=? AND cartItems.id IN (?)`;
-    conn.query(sql, [authorization.id, selected], // [user_id, ...selected] => [user_id, 1, 3]의 형태로 들어감
-        (err, results) => {
-            if(err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
-            }
+        WHERE user_id=?`;
+        let values = [authorization.id]
 
-            return res.status(StatusCodes.OK).json(results);
+        if (selected) { // 주문서 작성 시 '선택한 장바구니 목록 조회'
+            sql +=  ` AND cartItems.id IN (?)`;
+            values.push(selected);
         }
-    )}
+
+        conn.query(sql, values, 
+            (err, results) => {
+                if(err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+
+                return res.status(StatusCodes.OK).json(results);
+            }
+        )
+    }
 };
 
 const removeCartItem = (req, res) => {
-    const cartItemId = req.params.id
+    let authorization = ensureAuthorization(req, res);
 
-    let sql = `DELETE FROM cartItems WHERE id = ?`;
-    conn.query(sql, cartItemId,
-        (err, results) => {
-            if(err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
+    if(authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else {
+        const cartItemId = req.params.id
+
+        let sql = `DELETE FROM cartItems WHERE id = ?`;
+        conn.query(sql, cartItemId,
+            (err, results) => {
+                if(err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+
+                return res.status(StatusCodes.OK).json(results);
             }
-
-            return res.status(StatusCodes.OK).json(results);
-        }
-    )
-};
-
-function ensureAuthorization(req, res) {
-
-    try {
-        let receivedJwt = req.headers["authorization"];
-        console.log("recieved jwt : ", receivedJwt);
-    
-        let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-        console.log(decodedJwt);
-    
-        return decodedJwt;
-    } catch (err) {
-        console.log(err.name);
-        console.log(err.message);
-
-        return err;
+        )
     }
-}
+};
 
 module.exports = {addToCart, getCartItems, removeCartItem}
